Adjustment of shared.sd_upscalers on api.py
The format that was being used by api.py was not matching the one used in modules.realesrgan_model, so I imported the appropriate list structure and then changed shared.sd_upscalers in api.py to use the same format.

There is a field on the new objects called `location` which I did not change value of because I was not sure what to set it to. But the other change is working.

I did need to ctrl+c once on Windows to get it to work properly. It didn't actually close the running program, and after hitting ctrl+c once it then started generating images properly, for repeated requests.

Resolved error:
File "E:\StableDiffusion\imperator-maximus\stable-diffusion-webui\modules\realesrgan_model.py", line 58, in setup_realesrgan
    modules.shared.sd_upscalers.append(UpscalerRealESRGAN(model.netscale, i))
AttributeError: 'dict' object has no attribute 'append